+++ R/az_15min.R
@@ -1,7 +1,8 @@
 #' Retrieve 15-minute Weather Data from AZMet
 #'
 #' Retrieves 15-minute data from the AZMet (Arizona Meteorological Network) API.
-#' For a list of weather stations and their locations see [station_info].
+#' For a list of weather stations and their locations see [station_info], or
+#' visit https://azmet.arizona.edu/about.
 #'
 #' @param station_id Station ID can be supplied as numeric vector (e.g.
 #'   `station_id = c(8, 37)`) or as character vector with the prefix "az" and
@@ -16,7 +17,7 @@
 #'   Defaults to the current date and time if left blank and `start_date_time`
 #'   is specified.
 #' @details If neither `start_date_time` nor `end_date_time` are supplied, the
-#'   most recent date-time of data will be returned. If only `start_date_time`
+#'   most recent datetime of data will be returned. If only `start_date_time`
 #'   is supplied, then `end_date_time` defaults to the current time. Supplying
 #'   only `end_date_time` will result in an error.
 #' @note If `station_id` is supplied as a vector, multiple successive calls to
@@ -24,9 +25,9 @@
 #'   the stations by leaving `station_id` blank and subsetting the resulting
 #'   dataframe. Only the most recent 48 hours of 15-minute data are stored in
 #'   the AZMet API.
-#' @return a tibble. For units and other metadata, see
+#' @return A tibble. For units and other metadata, see
 #'   <https://azmet.arizona.edu/about>
-#' @seealso [az_daily()], [az_heat()], [az_hourly()], [az_lw15min()]
+#' @seealso [az_daily()], [az_heat()], [az_hourly()], [az_lw15min()], [az_lwdaily()]
 #' @source <https://azmet.arizona.edu/>
 #' @importFrom rlang .data
 #' @export
@@ -71,7 +72,7 @@ az_15min <- function(station_id = NULL, start_date_time = NULL, end_date_time = 
   # Query API ------------------------------------------------------------------
 
   if (is.null(start_date_time) & is.null(end_date_time)) {
-    message("Querying most recent date-time of 15-minute data ...")
+    message("Querying most recent datetime of 15-minute data ...")
   } else {
     message(
       "Querying data from ", format(params$start, "%Y-%m-%d %H:%M:%S")," through ", format(params$end, "%Y-%m-%d %H:%M:%S"), " ..."
@@ -110,7 +111,7 @@ az_15min <- function(station_id = NULL, start_date_time = NULL, end_date_time = 
   }
 
   if (nrow(out) == 0) {
-    warning("No data retrieved from API.")
+    warning("No data retrieved from API")
     # Return 0x0 tibble
     return(tibble::tibble())
   }